requirements.txt
flet
google-generativeai
beautifulsoup4
requests
